Refactor and enhance machining operations and post-processing scripts
- Updated DrillOp.py to import baseOp and CoolantMode from the new Op.BaseOp and utils modules, respectively. Removed redundant property definitions for FeedRate, SpindleSpeed, and CoolantMode.
- Modified OpContournage.py and OpSurfacage.py to align with the new import structure and improved G-code generation in Surfacage.
- Introduced a new PathOp.py file to handle path operations, including G-code generation and animation.
- Created utils.py to define a global CoolantMode enumeration for better management of coolant settings.
- Refactored ITnc530.py and Siemens828.py post-processor scripts to inherit from BasePostPro, enhancing code organization and maintainability.
- Updated ToolTaskPannel.py to integrate BQuantitySpinBox for better handling of tool parameters like Speed and Feed.
- Adjusted BQuantitySpinBox.py to improve property handling and widget updates, ensuring better integration with FreeCAD's property system.
- Updated test4.FCStd to reflect changes in the project structure and functionality.
