+++ Op/utils.py
@@ -1,0 +1,3 @@
+from enum import Enum
+
+CoolantMode = ["Off", "Flood", "Mist"]
